// types.ts
/** @format */

// types.ts

export interface EnrichedBreed {
  id: number;
  name: string;
  description: string;
  temperament: string;
  life_span: string;
  origin: string;
  weight: string;
  height: string;
  bred_for: string;
  breed_group: string;
  image_url: string;
  wikipedia_url: string;
}